minor: refactoring
