+++ crates/sandbox/src/deploy.rs
@@ -1,21 +1,17 @@
 use std::sync::Arc;
-use ethers::abi::Tokenize;
-use ethers::contract::ContractInstance;
-
 use starknet_core_contract_client::clients::{StarknetSovereignContractClient, StarknetEthBridgeContractClient, StarknetTokenBridgeContractClient, StarkgateManagerContractClient, StarkgateRegistryContractClient, DaiERC20ContractClient};
 
-use crate::{deploy_contract, Error, LocalWalletSignerMiddleware};
+use crate::{deploy_contract, deploy_contract_behind_unsafe_proxy, Error, LocalWalletSignerMiddleware};
 
 const STARKNET_SOVEREIGN: &str = include_str!("../artifacts/Starknet.json");
 
-const STARKNET_ETH_BRIDGE: &str = include_str!("../artifacts/StarknetEthBridge.json");
+const STARKNET_ETH_BRIDGE: &str = include_str!("../artifacts/StarknetLegacyBridge.json");
 
 const STARKGATE_MANAGER: &str = include_str!("../artifacts/StarkgateManager.json");
 const STARKGATE_REGISTRY: &str = include_str!("../artifacts/StarkgateRegistry.json");
 const STARKNET_TOKEN_BRIDGE: &str = include_str!("../artifacts/StarknetTokenBridge.json");
-const ERC20_TOKEN: &str = include_str!("../artifacts/DaiERC20.json");
+const DAI_ERC20_TOKEN: &str = include_str!("../artifacts/DaiERC20Token.json");
 
-const UNSAFE_PROXY: &str = include_str!("../artifacts/UnsafeProxy.json");
 
 /// Deploy Starknet sovereign contract and unsafe proxy for it.
 /// Cached forge artifacts are used for deployment, make sure they are up to date.
@@ -83,31 +79,12 @@ pub async fn deploy_dai_erc20_behind_unsafe_proxy(
     client: Arc<LocalWalletSignerMiddleware>
 ) -> Result<DaiERC20ContractClient, Error> {
     // Deploy the Dai ERC20 Token contract (no explicit constructor)
-    let contract = deploy_contract(client.clone(), ERC20_TOKEN, ()).await?;
+    let contract = deploy_contract(client.clone(), DAI_ERC20_TOKEN, ()).await?;
 
     Ok(DaiERC20ContractClient::new(
         contract.address(),
         client.clone(),
     ))
-}
-
-
-pub async fn deploy_contract_behind_unsafe_proxy<T: Tokenize>(
-    client: Arc<LocalWalletSignerMiddleware>,
-    contract_path: &str,
-    constructor_args: T,
-) -> Result<ContractInstance<Arc<LocalWalletSignerMiddleware>, LocalWalletSignerMiddleware>, Error> {
-    let contract = deploy_contract(client.clone(), contract_path, constructor_args).await?;
-
-    // Once we know the main contract address (implementation address)
-    // we can deploy and initialize our delegate proxy.
-    // NOTE that real world proxies typically allow changing the implementation
-    // address dynamically (this is basically how upgrades work). In our case,
-    // for simplicity, the proxy is initialized only once during the deployment.
-    let proxy_contract =
-        deploy_contract(client.clone(), UNSAFE_PROXY, contract.address()).await?;
-
-    return Ok(proxy_contract);
 }
 
 #[cfg(test)]
@@ -119,7 +96,7 @@ mod tests {
             CoreContractInitData, OperatorTrait, ProxyInitializeData, ProxySupportTrait,
             StarknetSovereignContractTrait,
         },
-        StarknetCoreContractClient,
+        StarknetContractClient,
     };
 
     #[tokio::test]
